set /content as Hugo module and import updated Contributing guidelines
Set up to use content as a Hugo module and automatically update it on every change in the site repository. Added details imported from old project's Readme as well. This will allow contributors to get the content and the site to test locally and submit PRs with review apps on GitHub, while we keep the existing git history from /content.