Simplify PR template
Former template was too detailed and not suited for quick changes like the changelog.

Furthermore, already pre-filled templates seem to reduce author's verbosity. Since there isn't any way of enforcing the filling of a template on PR's like on issues on GitHub, if any additional information is needed it should be requested by the reviewer.