Hugo version setup (#366)
This chore configuration sets a minimum (`0.133.0`) and a maximum
(`0.133.1`) Hugo version. The minimum is based on potential breaking
changes and the maximum is the last tested one.

This solves local errors depending on the installed Hugo version
installed across contributors machines. If the installed version is out
of scope, the following warning will print:

```shell
WARN  Module "github.com/CleverCloud/documentation" is not compatible with this Hugo version: 0.133.0/0.133.1
```

## Reviewers
_Who should review these changes?_ @Kirbeerus could you tell me if this
change would be helpful enough for this kind of errors?